backend/requirements.txt
fastapi
uvicorn[standard]

sqlalchemy

pydantic[email]

requests

python-jose[cryptography]
bcrypt
python-multipart

python-dotenv